Fix missing subtitles in some cases.
We were passing subtitles back from decoders to SubtitleDecoder
using dcp::SubtitleStrings and relying on their storage of time
to know when the subtitles were.  These times are quantised (by
the use of dcp::SubtitleString) and then compared with unquantised
times (kept as ContentTime) in the main checking loop in
SubtitleDecoder::get().

Fix this by storing periods as ContentTimePeriod as well as
in the dcp::SubtitleStrings.