Avoid decoding other packets when looking for subs.
The "accumulation" of, for example, video data when we are looking
for audio data is an *optimisation* to reduce the number of seeks.
It should not be necessary for correctness (the output should be right
even if we never kept anything except what we were looking for).

Doing this accumulation is not always an optimisation; sometimes not
doing it is better.  Avoiding it when going back for subtitles is one
of these cases.