Remove caching of old ImageDecoder objects.
This breaks things when there is a 3D ImageContent.  When
you change the video frame type on this content the view
does not update because the re-used ImageDecoder recycles
the same video without noticing that the frame type has changed.

I guess this is sort of `because' the video frame type is used
in VideoDecoder::give, which sets up the cache.

Unfortunately I can't remember the case which the caching
of ImageDecoders was meant to speed up.  Maybe this will
now become apparent.