Optimization for the referenced video case.
With referenced video from a DCP decoder, no video will ever
be fetched from the decoder.  Hence the code to discard given video
will be activated after _decoded builds up to the magic size.

Before this commit the code would attempt to fill with black up to
given frame N (with N very large) from the last frame in _decoded when
_decoded had been trimmed.  This would result in exponential growth
in execution time for the VideoDecoder::give() path.