Fix missing subtitle in some cases.
With this timeline:
-------> t
SUB1   SUB2    X

We might seek to X, then pass().  Before this change, SUB1 would have been emitted by
the call to the subtitle decoder, then we'd have emitted some black for X.  This would
lose SUB2.  Now we keep pass()ing and don't emit X until it's the earliest thing
(in the same way that the main decision of what to pass() works).