Potential fix for #1568; we calculate the remaining DCPTime but then convert it back to resampled content frames using the DCP rate rather than the resampled content rate, which seems wrong.
If we want to go from metric time to frames we surely have to use
the frame rate of the thing we are working with (not the frame rate
which that thing will be played back at).

Backported from 26c62598730d1d32333bfab0d5f463b90d26ee4d in v2.15.x.