Correct ext2 block size to 4096.
It seems that this is the typical Linux value; it allows large files
and so on.  But more importantly, it appears that this change turns
the lwext4 library code from taking an hour+ to do a USB stick format on mac
(and maybe other platforms) to taking under 3 minutes.

The underlying reason for this is not clear.