This AtomicityChecker was introduced a long time ago for reasons that are unclear.
This commit removes it because it was causing a bug whereby the
_suspended count on Butler would not go back down to 0 due to signals
being blocked by the AtomicityChecker.

I can see two possibilities:

1. AtomicityChecker is a band-aid for some deeper problem, in which
case other bugs will become apparent.  Perhaps it relates to signals
being sent before the butler/player/playlist are connected to content.

2. AtomicityChecker was correct but was not identifiying signals
from different pieces of content correctly; so the correct fix would
be to add the content pointer to the _checker.send() call and block
based also on that.