Fix deadlock when changing CPL in the player (#1827).
TextContent::set_dcp_track can end up requesting a view update, which
involves calls to methods in Content which lock the Content::_mutex.
Do these calls without a lock on that mutex held.

Also, it looks like we would append to texts on every call to
examine().  Fix that so that we replace the texts list on each
examine() call.