Coalesce player changes around CPL change.
If multiple player changes happen, causing the player to be suspended,
there are problems if one of those changes causes a FilmViewer:seek.
In there, we wait for the seek to finish, but it will never finish
because of the suspension.  Hence it's important that we only trigger
the seek once the suspension is over.

Fixes #1905.