Fix incorrectly-timed emission of silence padding causing buffer fill (#2217).
On initialisation or after seek we insert silence corresponding to
a positive delay in an audio stream.  Previously this inserted silence
was done at time 0, so that after a seek to time T the silent frames
would come out of the audio merger at time 0 and then the player would
fill the space up to time T with silence.  If T was far enough along
this would fill the audio buffers without there being any video.