Fix writing of planar audio to export buffers in some cases (#2223).
Previously we did not take into account alignment padding that can
exist with FLTP samples; each channel's block of samples can have
a gap between.  This doesn't happen with the normal 1024 sample blocks;
only when flushing at the end, when we write shorter blocks.

Not doing this right meant that we were passing uninitialised memory
to aacenc which sometimes responded by returning an EINVAL due to one
of its internal calculations returning a NaN.