Fix strange multi-selection behaviour on Linux (#2269).
It seems that all of a sudden the behaviour of wxListCtrl's selection
is strange (on Linux at least).  If a list item is selected by
some other action (e.g. selection in the timeline) and then you click
another directly in the list, both are selected.

This "fix" seems to work on Linux at least - also setting
wxLIST_STATE_FOCUSED as well as wxLIST_STATE_SELECTED.  I don't know
why this works.