Let config/cinemas/dkdm_recipients take care of their own recovery.
This should mean that if one fails to load, but the others are OK,
we don't re-write everything.  I hope this has no bad side-effects
that I didn't think of.

It's particularly helpful in the case of #2319 where a failure to
load cinemas.xml (which was being created by another tool) caused
the config to be re-written, breaking all existing DKDMs.