Remove scrolling in AudioMappingView.
In the audio panel it was a bit weird to have part of the panel
scrollable when the whole video panel is.  Now both video and audio
panels scroll.  You lose the "always there" view of the audio processing
summary but I think this is outweighed by the scrolling simplicity.

I also suspect it may make things behave better on "short" screens.