Use CURLOPT_NOSIGNAL to fix crashes on update check (#2495).
Without this, it's possible to come back from curl_easy_perform
in a different thread to the one that it was called from (in the case
of an error).

As I understand it, this happens because a signal handler is called
(from a different thread) and CURLOPT_NOSIGNAL stops this from
happening.