Default to writing 8 audio channels in DCPs (#2443).
It seems likely that 8 channels is slightly less likely to trip validation
warnings from some distributors, and so it seems worth accepting the potential
waste of 2 channels.  Users can always set this back to 6 again (once,
and it will be saved) and those upgrading from old versions won't be affected
anyway.