Make sure 2-channel MXFs don't get extra channel descriptors (#2631).
With all the EasyDCP tests done in bug #2487 we didn't look at stereo
MXFs as it appeared fairly early on that EasyDCP would reject them all.

In #2631 it's pointed out that it seems to make no sense to have descriptors
for channels which aren't there.  I've still got no docs for any of this,
but let's try fixing that.