Don't add a font with an empty ID for ccaps either (#2762).
Previously this was all changed so that for subs we add the default
font in a cleaner way (in db22f81ccce9e1a5f205e6d8b3c0631fc039a173).

I can't see why we shouldn't treat ccaps the same way, and indeed
not doing it causes an error as shown by the included test.