Fix hanging/overlapping dvb subtitles (#2792).
This reverts a change made in
8ca6fd6d97e6d42492afddb655fa85130946853c
"Fix doubled subtitles if subtitle stop times are specified."

That change breaks the case where a subtitle _does_ have a stop time,
but it's wrong (30s from the start time) and we want the next subtitle
to clear the previous one.

I can't now see how reverting this could cause doubled subtitles,
so maybe that problem wlil come back.  At least now there's a test
for #2792.