Don't emit video until film's end if we are seeking.
Otherwise if we seek to the frame just before the end of the DCP
we'll call emit_video_until() with _next_video unset, causing that
method to emit video for the whole film.

The point of this emit_video_until() call is only to pad a space
between the end of the video content and the end of the film.