Hopefully fix occasional hang in j2k_encoder_deadlock_test.
Previously too many frames were queued for encoding, which AFAICS meant
that (if the CPU was busy) we would get to the point where too many frames
were in the encoder queue, so that we blocked waiting for it to clear,
and then simultaneously too many frames were in the writer queue, which
(in this test) would never clear.

At this point we would be backed up waiting for Writer::write() to happen
in J2KEncoder::encoder_thread() so that the encoder queue could be cleared,
but nobody is calling Writer::write().