Fix multi-selection of content.
In order to add ellipsizing of text we're using this "mode" of wxListCtrl where
`OnGetItemAttr` is called to get information about an item.  This was intended
to support very long lists, so that the whole list does not have to be given
to wxListCtrl.

However, in this mode, multiple selections (e.g. shift-click) are just not
reported.  There's a 13-year-old wxWidgets bug to fix this:

https://github.com/wxWidgets/wxWidgets/issues/4541

and there is mentioned the workaround applied here - listen to the focused
event.  We have to delay it because on macOS (at least) the selection state
is not updated until after the focused event.