Fix doubled subtitles (again).
This was re-introduced when
94618a724124cbf5fe9f0b47a3fdce601fcd5581
reverted a previous attempt at a fix.

At the time I couldn't understand the doubled-subtitles problem,
but it's apparent in the test introduced in the next commit.

This is another attempt to fix it by only sending a "stop" for
a subtitle if we didn't already stop the subtitle because the
next one arrived.