Remove believed pointless pre-storage of file state.
I think I previously misunderstood what was going on here and thought
the problem was that doing set_cinemas_file() created a sqlite3 file.  I
now think that the problem is rather that we call Config::instance()
early in dcpomatic.cc, then ::drop().  This creates the sqlite3 file,
sets its path in the Config object but then deletes the object without
saving it.  When we create the Config object "for real" we don't reset
the path from XML to sqlite3.

The part of the previous fix that unconditionally sets the config path
seems to be the right one.