Fix macOS locale handling.
The docs for wxLocale say it's broken on macOS, and it seems to cause
problems when e.g. de_DE is configured in DoM and the computer is set to
de_DE locale (and language) - then you can't enter , in wxSpinCtrlDouble
and if you type a . and then use the up/down buttons it all goes wrong.

It doesn't seem possible to set a configured locale in DoM, but you
probably don't want that anyway.  wxUILocale::UseDefault() appears to
magically fix everything.

Unfortunately this is not available in the version of wxWidgets we use
for 10.8 and 10.9, and we can't go higher with wxWidgets on those macOS
versions.  So maybe localisation on 10.{8,9} will be broken, and stay
broken.