Don't use an optional for the Grok config object.
I'm not sure why I did it like this in the first place - perhaps so that
if the API endpoint changes there aren't so many old values stuck in
config files all over the place?  Anyway, it seems cleaner to do it like
this, as it's how we handle the other config.