Remove Ubuntu 22.04 test reference hacks.
Before the previous commit Ubuntu 22.04 would give slightly different
results to 24.04 when burning subtitles.  I think this was because
DoM was erroneously using the system copy of Liberation Sans, which
I presume was changed.

After the previous commit we are always using the in-tree Liberation
Sans, which means that we can update the test references for Ubuntu
24.04 and use them also for 22.04.