New/improved pixel format decision when converting to XYZ.
For a long time we would keep XYZ12LE, if that's what we have, otherwise
ask FFmpeg to switch to RGB48LE.

Then in 1d5c211dadb9a9dc2318adce86ca9c31b367cabe I tried to fix the case
of an XYZ source mis-tagged as YUV.  I changed things so that with no
colour conversion we'd always ask FFmpeg to convert to XYZ.  This meant
that RGB sources with no colour conversion would get treatment by FFmpeg
due to the RGB -> XYZ switch.

Here we're going back to the more-or-less the "long time" behaviour when
there is a conversion (keep XYZ12LE but otherwise convert to RGB48).

When there's no conversion, keep RGB (to avoid the FFmpeg conversion
from RGB -> XYZ) but convert everything else to XYZ.